move webhook tests (#36053)
