+++ src/webhooks/tests/api.js
@@ -1,8 +1,8 @@
-import { get } from '../helpers/e2etest.js'
+import { get } from '../../../tests/helpers/e2etest.js'
 import {
   SURROGATE_ENUMS,
   makeLanguageSurrogateKey,
-} from '../../middleware/set-fastly-surrogate-key.js'
+} from '../../../middleware/set-fastly-surrogate-key.js'
 import { describe, expect } from '@jest/globals'
 
 describe('webhooks v1 middleware', () => {
